Fix CRUD Delete / GET for character/campaign + seperate components/UserProfileForm
