+++ client/src/pages/CampaignDetail.js
@@ -1,15 +1,19 @@
 import React from 'react';
 
-const CampaignDetail = ({ id, name, description, isActive, handlePatchCampaign, handleDeleteCampaign }) => {
-    return (
-        <li key={id}>
-            <span>Name: {name}</span>
-            <span>Description: {description}</span>
-            <span>{`Is Active: ${isActive}`}</span>
-            <button onClick={() => handlePatchCampaign(id)}>Update</button>
-            <button onClick={() => handleDeleteCampaign(id)}>Delete</button>
-        </li>
-    );
+const CampaignDetail = ({ id, name, description, characters, handlePatchCampaign , handleDeleteCampaign}) => {
+
+  console.log(characters)
+  return (
+      <li>
+          <span>Name: {name}</span>
+          <span>Description: {description}</span>
+          <span>Active Characters in the campaign: 
+              {characters.map(characterCampaign => characterCampaign.character).join(', ')}
+            </span>
+          <button onClick={() => handlePatchCampaign(id)}>Update</button>
+          <button onClick={() => handleDeleteCampaign(id)}>Delete</button>
+      </li>
+    )
 };
 
 export default CampaignDetail;
