+++ client/src/pages/CharacterDetail.js
@@ -1,11 +1,11 @@
 import React from 'react';
 
-const CharacterDetail = ({ id, name, description, isActive, handlePatchCharacter, handleDeleteCharacter }) => {
+const CharacterDetail = ({ id, name, description, alive, handlePatchCharacter, handleDeleteCharacter }) => {
     return (
         <li key={id}>
             <span>Name: {name}</span>
             <span>Description: {description}</span>
-            <span>{`Is Active: ${isActive}`}</span>
+            <span>{`Is Active: ${alive}`}</span>
             <button onClick={() => handlePatchCharacter(id)}>Update</button>
             <button onClick={() => handleDeleteCharacter(id)}>Delete</button>
         </li>
